pyproject.toml
[tool.poetry]
name = "esce"
version = "0.1.2"
description = "Tools for eastimating and extrapolating learning curves."
authors = ["Marc-Andre Schulz <marc-andre.schulz@charite.de>"]
license = "GPLv3" 

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.22"  
scikit-learn = "^1.2"  
pandas = "^1.5"
scipy = "^1.9"
imbalanced-learn = "^0.11"
plotly = "^5.8"
kaleido = "0.2.1"  
pyyaml = "^6.0"
snakemake = "^7.30"

[tool.poetry.dev-dependencies]
pytest = "^7.0"
pytest-cov = "^4.0"
black = "^22.0"
flake8 = "^5.0"
sphinx = "^6.0"
sphinx_rtd_theme = "^1.0"
ruff = "^0.0.277"

[tool.pytest.ini_options]
testpaths = "tests"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.ruff]
select = ["E", "F", "I", "NPY", "PL", "PD", "SIM", "Q", "PT", "C4", "A", "B", "FBT", "UP" ]
ignore = ["PLR0913"]
fixable = ["A", "B", "C", "D", "E", "F", "G", "I", "N", "Q", "S", "T", "W", "ANN", "ARG", "BLE", "COM", "DJ", "DTZ", "EM", "ERA", "EXE", "FBT", "ICN", "INP", "ISC", "NPY", "PD", "PGH", "PIE", "PL", "PT", "PTH", "PYI", "RET", "RSE", "RUF", "SIM", "SLF", "TCH", "TID", "TRY", "UP", "YTT"]
unfixable = []
line-length = 88
# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
target-version = "py39"
include = ['esce/*.py', 'tests/*.py']

[tool.ruff.mccabe]
max-complexity = 10

[tool.black]
line-length = 88
target-version = ['py39']
include = 'esce\/.*\.pyi?$|tests\/.*\.pyi?$'
